+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+0.1.3 (2019-12-19)
+------------------
+
+* Added script osc_to_csv
+* Support for numpy array as timeseries
+
 0.1.2 (2019-07-04)
 ------------------